fix bugs and improve error handling
