+++ apps/backend/controllers/userControllers.go
@@ -10,48 +10,7 @@ import (
 	"net/http"
 
 	"github.com/gin-gonic/gin"
-	"golang.org/x/crypto/bcrypt"
 )
-
-func Signup(c *gin.Context) {
-	// Get the email/pass off req bodyu
-	var body struct {
-		Name     string
-		Email    string
-		Password string
-	}
-
-	if c.Bind(&body) != nil {
-		c.JSON(http.StatusBadGateway, gin.H{
-			"error": "Failed to read body",
-		})
-		return
-	}
-
-	// hash the password
-	hash, err := bcrypt.GenerateFromPassword([]byte(body.Password), 10)
-
-	if err != nil {
-		c.JSON(http.StatusBadGateway, gin.H{
-			"error": "Failed to hash password",
-		})
-		return
-	}
-
-	// Create the user
-	user := models.User{Name: body.Name, Email: body.Email, Password: string(hash), Role: "member"}
-	result := initializers.Db.Create(&user)
-
-	if result.Error != nil {
-		c.JSON(http.StatusBadGateway, gin.H{
-			"error": "Failed to create user",
-		})
-		return
-	}
-
-	// Respond
-	c.JSON(http.StatusCreated, user)
-}
 
 func GetAllUsers(c *gin.Context) {
 	email := c.Query("email")
