+++ apps/backend/initializers/loadEnvVariables.go
@@ -1,7 +1,7 @@
 package initializers
 
 import (
-	"log"
+	"fmt"
 
 	"github.com/joho/godotenv"
 )
@@ -10,6 +10,6 @@ func LoadEnvVariables() {
 	err := godotenv.Load()
 
 	if err != nil {
-		log.Fatal("Erro loading .env file")
+		fmt.Println("Erro loading .env file")
 	}
 }
